+++ pkg/cli/app.go
@@ -2,16 +2,13 @@ package cli
 
 import (
 	"context"
-	"fmt"
-	"io/ioutil"
-	"os"
+
+	"github.com/solo-io/build/pkg/ingest"
 
 	"github.com/solo-io/go-utils/clicore"
 	"k8s.io/helm/pkg/version"
 
 	"github.com/solo-io/go-utils/contextutils"
-
-	"github.com/solo-io/go-utils/protoutils"
 
 	v1 "github.com/solo-io/build/pkg/api/v1"
 	"github.com/solo-io/build/pkg/constants"
@@ -47,7 +44,7 @@ func Run() {
 func App(ctx context.Context, version string) *cobra.Command {
 	o := &Options{
 		Internal: Internal{ctx: ctx},
-		BuildRun: InitializeBuildRun(),
+		BuildRun: ingest.InitializeBuildRun(),
 	}
 	app := &cobra.Command{
 		Use:     "build",
@@ -129,89 +126,4 @@ func stringForBoolToEnv(b bool) string {
 		return constants.PrintEnvTrue
 	}
 	return constants.PrintEnvFalse
-}
-
-func getBuildRunConfigFromEnv(spec *v1.BuildSpec) v1.BuildRunConfig {
-	ev := &v1.BuildEnvVars{}
-	ev.BuildId = os.Getenv(constants.EnvBuildId)
-	ev.CommitSha = os.Getenv(constants.EnvCommitSha)
-	ev.TagVersion = os.Getenv(constants.EnvTagVersion)
-	cv := &v1.ComputedBuildVars{}
-	cv.Release = isRelease(ev)
-	cv.ImageTag = getImageTag(ev)
-	cv.ContainerPrefix = getContainerPrefix(cv.Release, spec.Config)
-	return v1.BuildRunConfig{
-		BuildEnvVars:      ev,
-		ComputedBuildVars: cv,
-	}
-}
-
-func getContainerPrefix(isRelease bool, config *v1.BuildConfig) string {
-	targetRegistry := config.ReleaseContainerRegistry
-	if !isRelease && config.TestContainerRegistry != nil {
-		targetRegistry = config.TestContainerRegistry
-	}
-	prefix := ""
-	if err := targetRegistry.GetPrefixFromContainerRegistry(&prefix); err != nil {
-		panic(err)
-	}
-	return prefix
-}
-
-func isRelease(ev *v1.BuildEnvVars) bool {
-	if ev.TagVersion == "" {
-		return false
-	}
-	return true
-}
-
-func getImageTag(ev *v1.BuildEnvVars) string {
-	tag := ev.BuildId
-	if isRelease(ev) {
-		tag = imageTagFromTaggedVersion(ev.TagVersion)
-	}
-	if tag == "" {
-		panic(fmt.Sprintf("must specify an image tag, none found for build env vars: %v", ev))
-	}
-	return tag
-}
-
-func imageTagFromTaggedVersion(tv string) string {
-	if len(tv) < 2 {
-		panic("must have at least two characters in TaggedVersion")
-	}
-	if tv[0] != 'v' {
-		panic(fmt.Sprintf("invalid tagged version: %v, must start with 'v'", tv))
-	}
-	return tv[1:]
-}
-
-func InitializeBuildRun() v1.BuildRun {
-	buildSpec := parseSpec()
-	buildRunConfig := getBuildRunConfigFromEnv(buildSpec)
-	return v1.BuildRun{
-		Spec:   buildSpec,
-		Config: &buildRunConfig,
-	}
-}
-
-// uses a config filename from env or default, in that order
-func parseSpec() *v1.BuildSpec {
-	// TODO - also get from flag (as first priority)
-	filename := ""
-	spec := &v1.BuildSpec{}
-	envFile := os.Getenv(constants.EnvVarConfigFileName)
-	if envFile != "" {
-		filename = envFile
-	} else {
-		filename = constants.DefaultConfigFileName
-	}
-	b, err := ioutil.ReadFile(filename)
-	if err != nil {
-		panic(err)
-	}
-	if err := protoutils.UnmarshalYaml(b, spec); err != nil {
-		panic(err)
-	}
-	return spec
 }
